requirements.txt
Flask==2.2.5
PyJWT==2.8.0
PyJWT==2.8.0
Werkzeug==2.2.3
